Restructure whole project.
