+++ example/ios/Flutter/flutter_export_environment.sh
@@ -1,10 +1,10 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=/Users/ayushpgupta/development/flutter_master/flutter"
-export "FLUTTER_APPLICATION_PATH=/Users/ayushpgupta/StudioProjects/pie_chart/example"
+export "FLUTTER_ROOT=/home/sahil/flutter"
+export "FLUTTER_APPLICATION_PATH=/home/sahil/IdeaProjects/pie_chart/example"
 export "FLUTTER_TARGET=lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"
 export "SYMROOT=${SOURCE_ROOT}/../build/ios"
-export "FLUTTER_FRAMEWORK_DIR=/Users/ayushpgupta/development/flutter_master/flutter/bin/cache/artifacts/engine/ios"
+export "FLUTTER_FRAMEWORK_DIR=/home/sahil/flutter/bin/cache/artifacts/engine/ios"
 export "FLUTTER_BUILD_NAME=1.0.0"
 export "FLUTTER_BUILD_NUMBER=1"
