+++ lib/src/legend.dart
@@ -1,0 +1,40 @@
+import 'package:flutter/material.dart';
+
+class Legend extends StatelessWidget {
+  Legend({
+    @required this.title,
+    @required this.color,
+    @required this.style,
+  });
+
+  final String title;
+  final Color color;
+  final TextStyle style;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.start,
+      mainAxisSize: MainAxisSize.min,
+      children: <Widget>[
+        Container(
+          margin: EdgeInsets.symmetric(vertical: 2.0),
+          height: 20.0,
+          width: 18.0,
+          color: color,
+        ),
+        SizedBox(
+          width: 8.0,
+        ),
+        Flexible(
+          fit: FlexFit.loose,
+          child: Text(
+            title,
+            style: style,
+            softWrap: true,
+          ),
+        )
+      ],
+    );
+  }
+}
