+++ lib/src/utils.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+const defaultChartValueStyle = TextStyle(
+  fontSize: 12,
+  fontWeight: FontWeight.bold,
+  color: Colors.black,
+);
+
+const defaultLegendStyle = TextStyle(
+  fontSize: 12,
+  fontWeight: FontWeight.bold,
+  color: Colors.black,
+);
+
+const List<Color> defaultColorList = [
+  Color(0xFFff7675),
+  Color(0xFF74b9ff),
+  Color(0xFF55efc4),
+  Color(0xFFffeaa7),
+  Color(0xFFa29bfe),
+  Color(0xFFfd79a8),
+  Color(0xFFe17055),
+  Color(0xFF00b894),
+];
